added all FXs typeworks

+++ include/audioRender/MixMachine/FAUST_VALS/Faust_interpolate.hpp
@@ -7,15 +7,15 @@
 class FaustInterpolate{
 public:
     int selectInterpolator;
-    int v1;
-    int v2;
-    int v3;
-    int v4;
-    int v5;
-    int v6;
-    int v7;
-    int v8;
-    int vZero;
+    float v1;
+    float v2;
+    float v3;
+    float v4;
+    float v5;
+    float v6;
+    float v7;
+    float v8;
+    float vZero;
     int frames;
     int timerActive;
